// frontend/src/app/generated/model/homelabPkgApisNetworkSiteV1SyncPolicy.ts
/**
 * Homelab API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { HomelabPkgApisNetworkSiteV1SyncPolicyMeta } from './homelabPkgApisNetworkSiteV1SyncPolicyMeta';
import { HomelabPkgApisNetworkSiteV1SyncPolicyStatus } from './homelabPkgApisNetworkSiteV1SyncPolicyStatus';

export interface HomelabPkgApisNetworkSiteV1SyncPolicy {
  generation?: number;
  id?: string;
  meta?: HomelabPkgApisNetworkSiteV1SyncPolicyMeta;
  resourceVersion?: number;
  status?: HomelabPkgApisNetworkSiteV1SyncPolicyStatus;
}